requirements.txt
pytest>=7.3.1
allure-pytest>=2.9.45
pytest-testconfig>=0.1.0
pytest-order>=1.0.1
selenium>=4.3.0
pytest-timeout>=2.1.0
opencv-python>=4.6.0
python-dotenv>=1.0.0
Pillow==9.4.0
numpy==1.23.5
